requirements.txt
annotated-types==0.5.0
async-timeout==4.0.3
backoff==2.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
coloredlogs==15.0.1
diskcache==5.6.1
future==0.18.3
humanfriendly==10.0
idna==3.4
numpy==1.25.2
pandas==2.1.4
pydantic==2.0
pydantic_core==2.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
requests-toolbelt==1.0.0
schedule==1.2.0
six==1.16.0
typing_extensions==4.7.1
tzdata==2023.4
urllib3==2.0.3
webex-bot==0.4.1
webexteamssdk==1.6.1
websockets==10.2
